+++ release/components/body/body-row.component.js
@@ -36,9 +36,9 @@ var DataTableBodyRowComponent = (function () {
             var cls = 'datatable-body-row';
             if (this.isSelected)
                 cls += ' active';
-            if (this.row.$$index % 2 !== 0)
+            if (this.rowIndex % 2 !== 0)
                 cls += ' datatable-row-odd';
-            if (this.row.$$index % 2 === 0)
+            if (this.rowIndex % 2 === 0)
                 cls += ' datatable-row-even';
             if (this.rowClass) {
                 var res = this.rowClass(this.row);
@@ -123,7 +123,7 @@ var DataTableBodyRowComponent = (function () {
     DataTableBodyRowComponent.decorators = [
         { type: core_1.Component, args: [{
                     selector: 'datatable-body-row',
-                    template: "\n    <div\n      *ngFor=\"let colGroup of columnsByPin; let i = index; trackBy: trackByGroups\"\n      class=\"datatable-row-{{colGroup.type}} datatable-row-group\"\n      [ngStyle]=\"stylesByGroup(colGroup.type)\">\n      <datatable-body-cell\n        *ngFor=\"let column of colGroup.columns; let ii = index; trackBy: columnTrackingFn\"\n        tabindex=\"-1\"\n        [row]=\"row\"\n        [expanded]=\"expanded\"\n        [isSelected]=\"isSelected\"\n        [column]=\"column\"\n        [rowHeight]=\"rowHeight\"\n        (activate)=\"onActivate($event, ii)\">\n      </datatable-body-cell>\n    </div>\n  "
+                    template: "\n    <div\n      *ngFor=\"let colGroup of columnsByPin; let i = index; trackBy: trackByGroups\"\n      class=\"datatable-row-{{colGroup.type}} datatable-row-group\"\n      [ngStyle]=\"stylesByGroup(colGroup.type)\">\n      <datatable-body-cell\n        *ngFor=\"let column of colGroup.columns; let ii = index; trackBy: columnTrackingFn\"\n        tabindex=\"-1\"\n        [row]=\"row\"\n        [expanded]=\"expanded\"\n        [isSelected]=\"isSelected\"\n        [rowIndex]=\"rowIndex\"\n        [column]=\"column\"\n        [rowHeight]=\"rowHeight\"\n        (activate)=\"onActivate($event, ii)\">\n      </datatable-body-cell>\n    </div>\n  "
                 },] },
     ];
     /** @nocollapse */
@@ -139,6 +139,7 @@ var DataTableBodyRowComponent = (function () {
         'row': [{ type: core_1.Input },],
         'offsetX': [{ type: core_1.Input },],
         'isSelected': [{ type: core_1.Input },],
+        'rowIndex': [{ type: core_1.Input },],
         'cssClass': [{ type: core_1.HostBinding, args: ['class',] },],
         'rowHeight': [{ type: core_1.HostBinding, args: ['style.height.px',] }, { type: core_1.Input },],
         'columnsTotalWidths': [{ type: core_1.HostBinding, args: ['style.width.px',] },],