+++ docs/changelog.md
@@ -1,4 +1,8 @@
 # Changelog
+## 10.0.1
+- Bug: Fix old references to `$$index` and `$$expanded`
+- Bug: Add `$$index` in cell templates as `rowIndex`
+
 ## 10.0.0
 - BREAKING: `$$expanded` is no longer on row of detail templates, use `let-expanded` to get pop now
 - Feature: Always center activated page on pager